+++ packages/playground-ui/src/domains/datasets/components/versions/index.ts
@@ -1,2 +1,1 @@
-export { DatasetVersionsPanel } from './dataset-versions-panel';
 export { DatasetItemVersionsPanel } from './dataset-item-versions-panel';
